+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to COMMIT will be documented in this file.
 
 ### Changed
 - Loading of nii data using (np.asanyarray( nii.dataobj )) in core and trk2dictionary
+- confidence_map from float64 to float32
 
 ## [1.4.6] - 2021-03-25